Include all sources in case of mixed languages
Include both c and cpp sources if both are used within same project.
Intellisense requires/prefers that the include paths & macros are
set to project settings but if the 2 languages have different settings,
intellisense may not work correctly. Intellisense does not work
properly if the settings are only per file