Increase sleep time to 0.5s when building single project
With VS22 building the whole solution often fails because a single
project and ninja project will start building at the same time. Increase
sleep time so that the single target project so that there is smaller
chance that the single project starts building. Still don't know how
to handle this more robustly